fix: user에 관계 없이 찜한 관광지 삭제되는 문제 해결

+++ src/main/java/com/example/letstrip/service/ScrapTourSpotService.java
@@ -54,8 +54,12 @@ public class ScrapTourSpotService {
         User user = userRepository.findById(userId).orElseThrow();
         if (user != null) {
             ScrapTourSpot scrapTourSpot = scrapTourSpotRepository.findByName(name);
-            scrapTourSpotRepository.delete(scrapTourSpot);
-            return scrapTourSpot;
+            if (scrapTourSpot.getUser() == user) {
+                scrapTourSpotRepository.delete(scrapTourSpot);
+                return scrapTourSpot;
+            } else {
+                throw new RuntimeException("Invalid user");
+            }
         } else {
             throw new RuntimeException("Invalid user");
         }
